services/ai-service/main.py
from fastapi import FastAPI, File, UploadFile, HTTPException
from fastapi.middleware.cors import CORSMiddleware
from fastapi.responses import JSONResponse
import cv2
import numpy as np
from PIL import Image
import io
import os
from typing import List, Dict, Any
import logging
from datetime import datetime

# Configure logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

app = FastAPI(
    title="StitchMe AI Service",
    description="AI-powered wound assessment and analysis service",
    version="1.0.0"
)

# Add CORS middleware
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],  # Configure properly for production
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

@app.get("/")
async def root():
    return {
        "service": "StitchMe AI Service",
        "version": "1.0.0",
        "status": "running",
        "timestamp": datetime.now().isoformat()
    }

@app.get("/health")
async def health_check():
    return {
        "status": "healthy",
        "timestamp": datetime.now().isostring(),
        "service": "ai-service"
    }

@app.post("/analyze-wound")
async def analyze_wound(
    file: UploadFile = File(...),
    lidar_data: str = None
):
    """
    Analyze wound from uploaded image and optional LiDAR data
    """
    try:
        # Validate file type
        if not file.content_type.startswith('image/'):
            raise HTTPException(status_code=400, detail="File must be an image")
        
        # Read image
        contents = await file.read()
        image = Image.open(io.BytesIO(contents))
        
        # Convert to OpenCV format
        cv_image = cv2.cvtColor(np.array(image), cv2.COLOR_RGB2BGR)
        
        # TODO: Implement actual AI model
        # For now, return mock analysis
        analysis_result = {
            "wound_detected": True,
            "confidence_score": 0.87,
            "severity": "moderate",
            "wound_area_cm2": 2.5,
            "detected_features": {
                "color_analysis": "reddish with some inflammation",
                "texture": "irregular surface",
                "edges": "well-defined borders"
            },
            "recommendations": [
                "Clean wound with saline solution",
                "Apply antiseptic",
                "Cover with sterile bandage",
                "Monitor for signs of infection"
            ],
            "requires_professional": False,
            "risk_assessment": "low",
            "treatment_recommendation": "Clean wound and apply antiseptic. Monitor healing progress."
        }
        
        logger.info(f"Analyzed wound image: {file.filename}")
        
        return {
            "success": True,
            "data": analysis_result,
            "message": "Wound analysis completed successfully"
        }
        
    except Exception as e:
        logger.error(f"Error analyzing wound: {str(e)}")
        raise HTTPException(status_code=500, detail=f"Analysis failed: {str(e)}")

@app.post("/analyze-vitals")
async def analyze_vitals(vital_data: Dict[str, Any]):
    """
    Analyze vital signs data
    """
    try:
        # TODO: Implement vital signs analysis
        # For now, return mock analysis
        vitals_analysis = {
            "heart_rate": vital_data.get("heart_rate", 72),
            "blood_pressure": vital_data.get("blood_pressure", "120/80"),
            "temperature": vital_data.get("temperature", 98.6),
            "oxygen_saturation": vital_data.get("oxygen_saturation", 98),
            "assessment": "Normal vital signs",
            "alerts": [],
            "recommendations": ["Continue monitoring"]
        }
        
        return {
            "success": True,
            "data": vitals_analysis,
            "message": "Vital signs analysis completed"
        }
        
    except Exception as e:
        logger.error(f"Error analyzing vitals: {str(e)}")
        raise HTTPException(status_code=500, detail=f"Vitals analysis failed: {str(e)}")

@app.post("/process-lidar")
async def process_lidar(lidar_data: Dict[str, Any]):
    """
    Process LiDAR data for 3D wound mapping
    """
    try:
        # TODO: Implement LiDAR processing
        # For now, return mock processing result
        lidar_result = {
            "depth_map_processed": True,
            "wound_volume_ml": 0.8,
            "surface_area_cm2": 2.5,
            "depth_measurements": {
                "max_depth_mm": 3.2,
                "avg_depth_mm": 1.8,
                "depth_distribution": "uniform"
            },
            "3d_model_url": "mock-3d-model-url",
            "measurements_accuracy": 0.95
        }
        
        return {
            "success": True,
            "data": lidar_result,
            "message": "LiDAR data processed successfully"
        }
        
    except Exception as e:
        logger.error(f"Error processing LiDAR: {str(e)}")
        raise HTTPException(status_code=500, detail=f"LiDAR processing failed: {str(e)}")

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(
        "main:app",
        host="0.0.0.0",
        port=8000,
        reload=True,
        log_level="info"
    )
