services/ai-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.4
torch==2.1.1
torchvision==0.16.1
tensorflow==2.15.0
scikit-learn==1.3.2
pandas==2.1.4
python-dotenv==1.0.0
pydantic==2.5.2
httpx==0.25.2
aiofiles==23.2.1
